refactor(react): keep topic compatibility inside the message bus
